+++ move_do.h
@@ -42,8 +42,8 @@ struct undo_t {
 
 extern void move_do_init();
 
-extern void move_do(board_t *board, int_fast32_t move, undo_t *undo);
-extern void move_undo(board_t *board, int_fast32_t move, const undo_t *undo);
+extern void move_do(board_t *board, uint_fast16_t move, undo_t *undo);
+extern void move_undo(board_t *board, uint_fast16_t move, const undo_t *undo);
 
 extern void move_do_null(board_t *board, undo_t *undo);
 extern void move_undo_null(board_t *board, const undo_t *undo);
