+++ trans.h
@@ -30,7 +30,7 @@ extern void trans_clear(trans_t *trans);
 
 extern void trans_inc_date(trans_t *trans);
 
-extern void trans_store(trans_t *trans, uint_fast64_t key, int_fast32_t move, int_fast32_t depth,
+extern void trans_store(trans_t *trans, uint_fast64_t key, uint_fast16_t move, int_fast32_t depth,
 						int_fast32_t min_value, int_fast32_t max_value);
 
 extern bool trans_retrieve(trans_t *trans, uint_fast64_t key, int_fast32_t *move, int_fast32_t *min_depth,
